% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.r
\name{bootstrapping}
\alias{bootstrapping}
\title{Boostrapping Data}
\usage{
bootstrapping(dat, n)
}
\arguments{
\item{dat}{a dataframe.}

\item{n}{a numeric value indicating the number of samples to take.}
}
\value{
It returns a datafame of length n of sampled rows from the original
data.
}
\description{
Samples data with replacement n amount of times.With n a user defined input.
}
