% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_cross_validation.R
\name{nested_cross_validation}
\alias{nested_cross_validation}
\title{nested cross validation}
\usage{
nested_cross_validation(
  dat,
  folds_den,
  mcRep,
  grid,
  ctrl,
  boot = F,
  m = NULL,
  model,
  group
)
}
\arguments{
\item{dat}{is a dataframe.}

\item{grid}{is a grid of values for hyperparameter tuning.}

\item{ctrl}{is the control variables for the model.}

\item{boot}{is the option to conduct bootstrapping on the training data, if desired. Default is false.}

\item{m}{is used to determing the number of bootstaps, if using bootstrapping.}

\item{model}{is the desired ML model to be used.}

\item{group}{is the clinical group to be modelled.}

\item{fold_den}{is the number of iterations of a MC simulation, if desired.}
}
\value{
it returns performance statistics for the NCV, and saves them to a csv file.
}
\description{
This funcion will model the adni data in a nested cross validation procedure,
which will be modelled using the specified ML technique model, with
hyperparameter tuning defined with values found in grid.
}
