% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_modelling_proc.r
\name{post_modelling_proc}
\alias{post_modelling_proc}
\title{Post modelling processing}
\usage{
post_modelling_proc(
  dat,
  totalnewprediction,
  totalprobabilities,
  clinGroup,
  mcPerf
)
}
\arguments{
\item{dat}{a dataframe.}

\item{totalnewprediction}{a matrix of predictions.}

\item{totalprobabilities}{a matrix of prediction probabilities.}

\item{clinGroup}{a string indicating the clinical group of interest.}

\item{mcPerf}{is the performance from the pervious iteration.}
}
\value{
it returns a vector of five summary statistics.
}
\description{
This function will work in the outer loop of the bootstrap and spit out
summary statistics for the model.
}
