% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.r
\name{preprocessing}
\alias{preprocessing}
\title{preprocessing from adni_slim data}
\usage{
preprocessing(dat, perc, clinGroup)
}
\arguments{
\item{dat}{is a dataframe}

\item{perc}{is the percentage of missing values, where if
column missing > perc, column is removed.}

\item{clinGroup}{is the clinical group desired to filter on.}
}
\value{
It returns a dataframe of the preprocessed data
}
\description{
preprocesses the adni_slim data by removing columns with perc missing values
and filters on the relevant clinical group clinGroup. Also dummies and records
location of missing values.
}
